Refactor the class selectors to element selectors
The current approach to centering the content in each page is through
using class selectors called .format and .recipe, which works but
becomes tedious since each element containing text needs to be given a
class HTML tag. This change will apply the formatting to elements and
give a better way to apply the centered, but left aligned text within
a block. A margin will also be applied in the block.