Update README.md to mention CSS
This is the second attempt to push a commit regarding the README.md
being updated. I pushed the initial commit but I realized that I
wanted to change the commit message because the previous one used
past tense and was not very descriptive. I learned how to undo a push
to the remote repository and also became aware that I should first
check that everyone with a clone of this repo won't push their local
repo that doesn't contain this revert.